+++ migrations/00004_create_tasks_edges.sql
@@ -2,9 +2,9 @@
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS  tasks_edges (
     id BIGSERIAL PRIMARY KEY,
-    pipeline_id bigserial NOT NULL REFERENCES pipelines(id) ON DELETE CASCADE,
-    from_node BIGSERIAL NOT NULL REFERENCES tasks(id) ON DELETE CASCADE,
-    to_node BIGSERIAL NOT NULL REFERENCES tasks(id) ON DELETE CASCADE
+    pipeline_id INTEGER NOT NULL REFERENCES pipelines(id) ON DELETE CASCADE,
+    from_node INTEGER NOT NULL REFERENCES tasks(id) ON DELETE CASCADE,
+    to_node INTEGER NOT NULL REFERENCES tasks(id) ON DELETE CASCADE
 );
 -- +goose StatementEnd
